Don't try to guess when we're done; instead, just busy-wait.
GirlFriday::WorkQueue#wait_for_empty polls every 1/10th of a second (or
so) for whether the queue is empty.  However, the work queue can become
transiently empty if

(1) all workers are working on tasks in the queue
(2) one worker finishes a page with no next page link
(3) therefore, nothing's left in the queue

In that case, #wait_for_empty's busy-wait loop will terminate when we
really didn't want it to.